requirements.txt
crewai==0.28.8
crewai-tools==0.1.7
langchain-openai==0.1.6
langchain-community==0.2.1
python-dotenv==1.0.1
PyPDF2==3.0.1
fpdf2==2.7.9
requests==2.31.0
beautifulsoup4==4.12.3
